chore : ディレクトリ構成の変更
