+++ src/pages/Counseling.vue
@@ -1,4 +1,3 @@
 <template>
     <h1>カウンセリングを予約するページ</h1>
 </template>
-CounselingPage
