+++ src/pages/School.vue
