--- src/pages/Staff/StaffCreatePage.vue
@@ -1,22 +1,0 @@
-<template>
-    <Breadcrumbs :pages="pages" />
-    <StaffForm :staff="staff" />
-</template>
-
-<script setup lang="ts">
-import Breadcrumbs from "../../components/Breadcrumbs.vue";
-import StaffForm from "../../components/Staff/StaffForm.vue";
-
-const pages = [
-    { name: "スタッフ一覧", href: "staffs", current: false },
-    { name: "スタッフ新規作成", href: "staffCreate", current: true },
-];
-
-const staff = {
-    firstName: "",
-    lastName: "",
-    firstNameKana: "",
-    lastNameKana: "",
-    email: "",
-};
-</script>
